requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-generativeai
openai
google-cloud-texttospeech
groq
pypdf
python-docx